Fixing advanced/custom-sort example
As far as I can tell, setting `attr` does not have an effect as to what is used for arguments `a` and `b`. 

i.e. setting `attr` to "name" will not set `a` to `collectionMember.name` for comparison purposes.

It's obviously easier to fix up the docs than to change the Sort internals...however, maybe it's better this way, to not make any use of an `attr` setting, because the common use-case for a custom Sort is to work with the member objects directly, for elaborate sorting procedures?
 
 
(It's possible that I just missed a huge step in the docs and source-code, but `a` and `b` were not affected by whatever I set `attr` to)